fix loading init and historical states
