+++ scripts/gas-refund-program/staking/2.0/AbstractStateTracker.ts
@@ -1,5 +1,4 @@
 import { assert } from 'ts-essentials';
-import { BlockInfo } from '../../../../src/lib/block-info';
 import {
   CHAIN_ID_GOERLI,
   CHAIN_ID_MAINNET,
@@ -24,49 +23,6 @@ export class AbstractStateTracker {
     );
   }
 
-  async loadHistoricalstatesWithinInterval({
-    startTimestamp,
-    endTimestamp,
-  }: {
-    startTimestamp: number;
-    endTimestamp: number;
-  }) {
-    await this.resolveBlockBoundary({ startTimestamp, endTimestamp });
-    await this.loadStates();
-  }
-
-  async resolveBlockBoundary({
-    startTimestamp,
-    endTimestamp,
-  }: {
-    startTimestamp: number;
-    endTimestamp: number;
-  }) {
-    const blockInfo = BlockInfo.getInstance(this.chainId);
-    const [_startBlock, _endBlock] = await Promise.all([
-      blockInfo.getBlockAfterTimeStamp(startTimestamp),
-      blockInfo.getBlockAfterTimeStamp(endTimestamp),
-    ]);
-
-    assert(
-      typeof _endBlock === 'number' && _endBlock > 0,
-      '_endBlock should be a number greater than 0',
-    );
-    assert(
-      typeof _startBlock === 'number' &&
-        _startBlock > 0 &&
-        _startBlock < _endBlock,
-      '_startBlock should be a number and 0 < _startBlock < endBlock',
-    );
-
-    this.setBlockTimeBoundary({
-      startTimestamp,
-      endTimestamp,
-      startBlock: _startBlock,
-      endBlock: _endBlock,
-    });
-  }
-
   getBlockTimeBoundary(): BlockTimeBoundary {
     return {
       startTimestamp: this.startTimestamp,
@@ -86,10 +42,6 @@ export class AbstractStateTracker {
     this.endTimestamp = endTimestamp;
     this.startBlock = startBlock;
     this.endBlock = endBlock;
-  }
-
-  protected loadStates() {
-    throw new Error('Implement on child class');
   }
 
   assertTimestampWithinLoadInterval(timestamp: number) {
